Refactor:refine module.
