+++ client/src/options/clientOptions.ts
@@ -4,47 +4,23 @@ import {
   workspace,
   WorkspaceFolder,
 } from "vscode"
+import { ExecuteCommandSignature } from "vscode-languageclient"
 import {
-  LanguageClient, LanguageClientOptions, ServerOptions,
+  LanguageClientOptions,
 } from "vscode-languageclient/node"
 
-import { executeCommand } from "./commands"
-
-const PLPGSQL_LANGUAGE_SERVER_SECTION = "plpgsqlLanguageServer"
-
-
-export function createLanguageClient(
-  serverOptions: ServerOptions, clientOptions: LanguageClientOptions,
-) {
-  return new LanguageClient(
-    PLPGSQL_LANGUAGE_SERVER_SECTION,
-    "PL/pgSQL Language Server",
-    serverOptions,
-    clientOptions,
-  )
-}
-
+export const PLPGSQL_LANGUAGE_SERVER_SECTION = "plpgsqlLanguageServer"
 
 export function makeLanguageClientOptions(
   workspaceFolder?: WorkspaceFolder,
 ): LanguageClientOptions {
   let documentSelector
   if (workspaceFolder === null) {
-    documentSelector = [
-      {
-        scheme: "untitled",
-        language: "postgres",
-      },
-    ]
+    documentSelector = [{ scheme: "untitled", language: "postgres" }]
   }
   else {
-    documentSelector = [
-      {
-        scheme: "file",
-        language: "postgres",
-        pattern: `${workspaceFolder.uri.fsPath}/**/*`,
-      },
-    ]
+    const pattern = `${workspaceFolder.uri.fsPath}/**/*`
+    documentSelector = [{ scheme: "file", language: "postgres", pattern }]
   }
   const outputChannel: OutputChannel = window.createOutputChannel(
     PLPGSQL_LANGUAGE_SERVER_SECTION,
@@ -62,5 +38,16 @@ export function makeLanguageClientOptions(
     middleware: {
       executeCommand,
     },
+  }
+}
+
+
+async function executeCommand(
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  command: string, args: any[], next: ExecuteCommandSignature,
+): Promise<void> {
+  if (command === "plpgsql-lsp.executeFileQuery") {
+    args.push(window.activeTextEditor.document.uri.toString())
+    next(command, args)
   }
 }
