+++ client/src/options/index.ts
@@ -1,0 +1,5 @@
+export {
+  makeLanguageClientOptions,
+  PLPGSQL_LANGUAGE_SERVER_SECTION,
+} from "./clientOptions"
+export { makeLanguageServerOptions } from "./serverOptions"
