+++ server/src/postgres/index.ts
@@ -1,0 +1,1 @@
+export { getPool, PostgresClient, PostgresPool, PostgresPoolMap } from "./pool"
