+++ server/src/services/hover.ts
@@ -6,7 +6,7 @@ import {
 } from "vscode-languageserver"
 import { TextDocument } from "vscode-languageserver-textdocument"
 
-import { PostgresPool } from "@/postgres/pool"
+import { PostgresPool } from "@/postgres"
 import {
   makeFunctionDefinitionText,
   queryFunctionDefinitions,
